move auto-generated typings, adapt .gitignore
