+++ .travis.yml
@@ -3,7 +3,9 @@ node_js: node
 
 script:
   # will run `npm ci` automatically here
+  # - npm test
   - npm run lint
+  - npm run codegen
   # TODO: activate once this issue is solved:
   # https://github.com/graphprotocol/graph-ts/issues/113
   # - npm run type-check
