+++ src/mappings/datatoken.ts
@@ -1,7 +1,7 @@
 import { BigInt, BigDecimal } from '@graphprotocol/graph-ts'
-import { OrderStarted, Transfer } from '../types/templates/DataToken/DataToken'
+import { OrderStarted, Transfer } from '../@types/templates/DataToken/DataToken'
 
-import { Datatoken, TokenBalance, TokenOrder } from '../types/schema'
+import { Datatoken, TokenBalance, TokenOrder } from '../@types/schema'
 import {
   tokenToDecimal,
   updateTokenBalance,
